+++ erp-system-backend/src/main/java/com/example/erp_system/service/ClientService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface ClientService {
     List<User> getAllClients();
-    User getClientById(Long id);
+    User getClientById(int id);
     User createClient(User client);
-    User updateClient(Long id, User clientDetails);
-    void deleteClient(Long id);
+    User updateClient(int id, User clientDetails);
+    void deleteClient(int id);
 }
