+++ erp-system-backend/src/main/java/com/example/erp_system/service/EmployerService.java
@@ -6,8 +6,8 @@ import java.util.List;
 
 public interface EmployerService {
     List<User> getAllEmployers();
-    User getEmployerById(Long id);
+    User getEmployerById(int id);
     User createEmployer(User employer);
-    User updateEmployer(Long id, User employerDetails);
-    void deleteEmployer(Long id);
+    User updateEmployer(int id, User employerDetails);
+    void deleteEmployer(int id);
 }
